+++ packages/web/src/components/controls/Switch/Components/index.ts
@@ -1,0 +1,2 @@
+export * from './StatefulSwitch'
+export * from './StaticSwitch'