+++ packages/web/src/components/controls/Switch/index.ts
@@ -1,0 +1,8 @@
+import { withStyleProps } from '@shallot-ui/next'
+
+import { StatefulSwitch } from './Components'
+import { getSwitchStyles } from './getStyles'
+
+export const Switch = withStyleProps(StatefulSwitch, getSwitchStyles)
+
+export * from './types'