+++ apps/web/CHANGELOG.md
@@ -1,5 +1,11 @@
 # web
 
+## 0.3.15
+
+### Patch Changes
+
+- @shallot-ui/prismjs@4.0.5
+
 ## 0.3.14
 
 ### Patch Changes